requirements.txt
streamlit
extra-streamlit-components
requests
pymongo
python-dotenv
pyjwt
faker
pytest
fuzzywuzzy